Izvorni kod/backend/requirements.txt
flask
boto3
opencv-python
Pillow
numpy
imutils
pytesseract